+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Highlight selected node in Node.writeImageFile, childScopeDepth is -1 by default
 - Instantiation of a scope requires a blue print
 - Scopes can now have key aliases
+- Rename writeDotFile into writeImageFile
 
 ### Fixed